Moving document category icons
