+++ client/app/reader/DocumentCategoryIcons.jsx
