+++ client/test/karma/components/DocumentCategoryIcons-test.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { expect } from 'chai';
 import { mount } from 'enzyme';
-import { DocumentCategoryIcons } from '../../../app/components/DocumentCategoryIcons';
+import { DocumentCategoryIcons } from '../../../app/reader/DocumentCategoryIcons';
 
 describe('DocumentCategoryIcons', () => {
   it('renders no icons when the doc is not in any categories', () => {
